pyproject.toml
[build-system]
requires = ["maturin>=1.9,<2"]
build-backend = "maturin"

[project]
name = "spectra"
version = "0.1.0"
description = "Realistic radar and communication waveform generation with PyTorch integration"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.24",
    "torch>=2.0",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
zarr = ["zarr>=2.16"]
all = ["pyyaml>=6.0", "zarr>=2.16"]
dev = [
    "maturin>=1.9,<2",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pyyaml>=6.0",
    "zarr>=2.16",
]

[tool.maturin]
manifest-path = "rust/Cargo.toml"
python-source = "python"
module-name = "spectra._rust"
bindings = "pyo3"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = ["--tb=short", "-v", "--strict-markers"]
markers = [
    "slow: marks tests as slow",
    "rust: marks tests that exercise Rust FFI directly",
]
